feat: init sever api with two classes `Server` and `Connection`
add two classes `Server` and `Connection` as a first step of the sockets api. the 'Server' class represent the PORT socket, while 'Connection' class represent each connection (socket) with address bound to PORT of 'Server'.

The class 'Server' contain a vector of type 'Connection' each socket created in the 'Server' class add a new 'Connection' class to the vector. Its constructor takes PORT number as parameter and create the socket and make ready for connection, it have a method 'connect()' that accept connection and save the new socket fd.

The 'Connection' class is the main holder of sockets file descriptors. It got methods 'read()/send()' that read/write from/to the class '_socketfd', data read on 'read()' to the 'std::string' attributes '_request' and written from the 'std::string' attributes '_response' upon a 'send()' call.